% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{check_keyring_entry_exists}
\alias{check_keyring_entry_exists}
\title{Checks whether a specified entry exists in the current system keyring}
\usage{
check_keyring_entry_exists(service_name, user_name)
}
\arguments{
\item{service_name}{keyring service_name entry to check}

\item{user_name}{keyring username entry to check}
}
\value{
Boolean indicating whether entry is in the keyring or not
}
\description{
Checks whether a specified entry exists in the current system keyring
}
\examples{
check_keyring_entry_exists('database', 'username')
}
