% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{check_needed_entries_exist}
\alias{check_needed_entries_exist}
\title{Checks every required keyring entry and whether they're defined in keyring}
\usage{
check_needed_entries_exist(keyring_dataset)
}
\arguments{
\item{keyring_dataset}{the keyring yaml data per standard spec}
}
\value{
list indicating whether the keyring entry exists or not
}
\description{
Checks every required keyring entry and whether they're defined in keyring
}
\examples{
keyring_block_list_one <- list("service" = "test", "username" = "hardcode")
keyring_block_list_two <- list("service" = "trial", "username" = "example")
keyring_data_format <- list("email_format" = "",
"keyring_entries" = list("secret_one" = keyring_block_list_one,
                         "secret_two" = keyring_block_list_two))
check_needed_entries_exist(keyring_data_format)
}
