% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{check_secret_entry_exists}
\alias{check_secret_entry_exists}
\title{checks both username and password entry exist for the keyring entry}
\usage{
check_secret_entry_exists(keyring_block)
}
\arguments{
\item{keyring_block}{a list capturing the info from the keyring yaml spec}
}
\value{
boolean indicate present of complete keyring entry
}
\description{
checks both username and password entry exist for the keyring entry
}
\examples{
keyring_block_list <- list("service" = "test", "username" = "hardcode")
check_secret_entry_exists(keyring_block_list)
}
