% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{construct_guide_message}
\alias{construct_guide_message}
\title{Builds an informative message to summarise missing entries and next step}
\usage{
construct_guide_message(keyring_table)
}
\arguments{
\item{keyring_table}{a tibble with one entry per keyring with service, user}
}
\value{
input table with column for guide message indicate missing entries.
}
\description{
Builds an informative message to summarise missing entries and next step
}
\examples{
# Creating a sample keyring table
example_table <- dplyr::tibble(
name = c("Key1", "Key2", "Key3"),
service = c("ServiceA", "ServiceB", "ServiceC"),
username = c("UserA", NA, "UserC"),
comment = c("Comment 1", "Comment 2", "Comment 3")
)
construct_guide_message(example_table)
}
