% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_entries.R
\name{convert_entry_to_uppercase}
\alias{convert_entry_to_uppercase}
\title{Takes a keyring entry and converts it to upper case}
\usage{
convert_entry_to_uppercase(entry_string)
}
\arguments{
\item{entry_string}{a string for a keyring entry}
}
\value{
an upper cases version of the keyring entry
}
\description{
Takes a keyring entry and converts it to upper case
}
\examples{
convert_entry_to_uppercase("lower")
}
