% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_entries.R
\name{extract_email_domain}
\alias{extract_email_domain}
\title{Pulls out the domain from an email format string}
\usage{
extract_email_domain(email_format)
}
\arguments{
\item{email_format}{a string specifying the required email address format}
}
\value{
the domain from an email format string
}
\description{
Pulls out the domain from an email format string
}
\examples{
specified_email_format <- "first_name.last_name@my_domain.com"
extract_email_domain(specified_email_format)
}
