% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_entries.R
\name{fill_special_entry}
\alias{fill_special_entry}
\title{Takes an input entry string and replaces certain special entries}
\usage{
fill_special_entry(candidate_string, email_format, email_inferer)
}
\arguments{
\item{candidate_string}{A string with possible special entries to replace}

\item{email_format}{a string specifying the required email address format}

\item{email_inferer}{a function to infer email address start}
}
\value{
a parse entry with any special values replaced with fixed string
}
\description{
Takes an input entry string and replaces certain special entries
}
\examples{
fill_special_entry(
  "{system_user}CustomEnding",
  "first_name.last_name@my_domain.com",
  function(x) {
    "hello"
  }
)
}
