% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_entries.R
\name{generic_email_inferer}
\alias{generic_email_inferer}
\title{Takes an email format, ignores it and returns user name for start of email}
\usage{
generic_email_inferer(email_format)
}
\arguments{
\item{email_format}{a string specifying the required email address format}
}
\value{
an inferred email prefix from the supplied format type
}
\description{
Takes an email format, ignores it and returns user name for start of email
}
\examples{
specified_email_format <- "first_name.last_name@my_domain.com"
generic_email_inferer(specified_email_format)
}
