% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{get_password_from_block}
\alias{get_password_from_block}
\title{Retrieves the password for an entry from the keyring matching current block}
\usage{
get_password_from_block(keyring_block)
}
\arguments{
\item{keyring_block}{a list capturing the info from the keyring yaml spec}
}
\value{
a string with the password captured in the keyring
}
\description{
Retrieves the password for an entry from the keyring matching current block
}
\examples{
\dontrun{
# Add a temporary keyring entry for testing
keyring::key_set_with_value("database", "username", "my_user")
keyring::key_set_with_value("database", "my_user", "my_pswd")

keyring_block_list_one <- list("service" = "database",
                               "username" = "my_user")
get_password_from_block(keyring_block_list_one)

# clean up: remove temp keyring entry
keyring::key_delete("database", "username")
keyring::key_delete("database", "my_user")
}
}
