% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{get_username_from_data}
\alias{get_username_from_data}
\title{Retrieves the username for an entry from the keyring dataset}
\usage{
get_username_from_data(keyring_dataset, entry_name)
}
\arguments{
\item{keyring_dataset}{the keyring yaml data per standard spec}

\item{entry_name}{name of the entry to get from the keyring dataset}
}
\value{
a string with username captured in the keyring
}
\description{
Retrieves the username for an entry from the keyring dataset
}
\examples{
\dontrun{
keyring::key_set_with_value("database", "username", "my_user")
keyring::key_set_with_value("database", "my_user", "my_pswd")
keyring::key_set_with_value("database2", "username", "my_user")
keyring::key_set_with_value("database2", "my_user", "my_pswd")

keyring_block_list_one <- list("service" = "database",
                               "username" = "my_user")
keyring_block_list_two <- list("service" = "database2",
                               "username" = "my_user")
keyring_data_format <- list("email_format" = "",
"keyring_entries" = list("secret_one" = keyring_block_list_one,
                         "secret_two" = keyring_block_list_two))
get_username_from_data(keyring_data_format, "secret_one")

# clean up: remove temp keyring entry
keyring::key_delete("database", "username")
keyring::key_delete("database", "my_user")
keyring::key_delete("database2", "username")
keyring::key_delete("database2", "my_user")
}
}
