% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_format.R
\name{keyring_yaml_to_df}
\alias{keyring_yaml_to_df}
\title{Unnests dataset entries into a tibble of required secret entries}
\usage{
keyring_yaml_to_df(yaml_data_as_list)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}
}
\value{
tibble containing one row for each secret required
}
\description{
Unnests dataset entries into a tibble of required secret entries
}
\examples{
yaml_data_example <- list("keyring_entries" = list(
list("name" = "test_one", my_row_data = 1),
list("name" = "test_two", my_row_data = 2),
list("name" = "test_three", my_row_data = 3)))
table_of_yaml_data <- keyring_yaml_to_df(yaml_data_example)
}
