% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_format.R
\name{load_keyring_yaml}
\alias{load_keyring_yaml}
\title{Loads in keyring yaml and try to process special entries}
\usage{
load_keyring_yaml(path_to_yaml_file, custom_email_processor)
}
\arguments{
\item{path_to_yaml_file}{string with path to yaml file}

\item{custom_email_processor}{function which can resolve email template}
}
\value{
a prepared yaml dataset for checking keyring entries
}
\description{
Loads in keyring yaml and try to process special entries
}
\examples{
local_folder_yaml <- system.file("extdata", "example_keyring_entries.yaml",
  package = "abKeyringR"
)
default_email_parser <- function(x) {
  "hello"
}
load_keyring_yaml(local_folder_yaml, default_email_parser)
}
