% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_entries.R
\name{load_username}
\alias{load_username}
\title{Gets system user name from system info}
\usage{
load_username()
}
\value{
The current users login name
}
\description{
Gets system user name from system info
}
\examples{
load_username()
}
