% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_format.R
\name{promote_keyring_entry_names}
\alias{promote_keyring_entry_names}
\title{Sets entry names for keyring entry to match their name field}
\usage{
promote_keyring_entry_names(yaml_data_as_list)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}
}
\value{
yaml keyring entries
}
\description{
Sets entry names for keyring entry to match their name field
}
\examples{
test_yaml <- list(
  "email_format" = "first_name.last_name@company.com",
  "keyring_entries" = list(
    list("name" = "keyring_one"),
    list("name" = "keyring_two")
  )
)
promote_keyring_entry_names(test_yaml)
}
