% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keyring.R
\name{raise_missing_entry_message}
\alias{raise_missing_entry_message}
\title{Raises an informative message to suggest what needs to be done.}
\usage{
raise_missing_entry_message(keyring_table)
}
\arguments{
\item{keyring_table}{a tibble with one entry per keyring with service, user}
}
\value{
input table only for missing entries + guide message and warning
}
\description{
Raises an informative message to suggest what needs to be done.
}
\examples{
# Creating a sample keyring table
example_table <- dplyr::tibble(
name = c("Key1", "Key2", "Key3"),
service = c("ServiceA", "ServiceB", "ServiceC"),
username = c("UserA", NA, "UserC"),
comment = c("Comment 1", "Comment 2", "Comment 3"),
entry_exists = c(FALSE, FALSE, TRUE)
)
raise_missing_entry_message(example_table)
}
