% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_format.R
\name{update_keyring_block}
\alias{update_keyring_block}
\title{Updates the username entry within a single keyring block}
\usage{
update_keyring_block(keyring_block, email_template, email_processor)
}
\arguments{
\item{keyring_block}{a single keyring entry from the yaml data}

\item{email_template}{a string specifying the required email address format}

\item{email_processor}{function which can resolve email template}
}
\value{
infers and updates username and returns keyring entry with update
}
\description{
Updates the username entry within a single keyring block
}
\examples{
email_start_inference <- function(x) {
  "hello"
}
specified_email_format <- "first_name.last_name@my_domain.com"
keyring_example_block <- list(
  "name" = "keyring_one",
  "username" = "harcoded_user"
)
update_keyring_block(
  keyring_example_block, specified_email_format,
  email_start_inference
)
}
