% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_format.R
\name{update_list_of_keyrings}
\alias{update_list_of_keyrings}
\title{Updates entries for every keyring in the yaml list}
\usage{
update_list_of_keyrings(yaml_data_as_list, email_template, email_processor)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}

\item{email_template}{a string specifying the required email address format}

\item{email_processor}{function which can resolve email template}
}
\value{
a yaml keyring entry with resolved usernames
}
\description{
Updates entries for every keyring in the yaml list
}
\examples{
simple_data_update <- function(dataset_block, dataset_defaults) {
  dataset_block$default <- "test_value"
  return(dataset_block)
}
test_yaml <- list(
  "default_script_path" = ".",
  "datasets_to_load" = list(
    list("name" = "dataset_one"),
    list("name" = "dataset_two")
  )
)
update_list_of_keyrings(test_yaml, simple_data_update)
}
